change name  PDM  in all files

+++ Core/Inc/PDM/pdm_init_IC.h
@@ -5,7 +5,7 @@
 
 
 ErrorStatus PDM_init_TMP1075( _PDM *pdm_ptr, I2C_TypeDef *I2Cx, uint8_t tmp1075_addr, uint8_t i2c_mux_addr, uint8_t i2c_mux_ch );
-ErrorStatus PDM_init_INA231( _PDM *pdm_ptr, I2C_TypeDef *I2Cx, uint8_t INA231_addr, float max_current, float Rshunt, uint8_t i2c_mux_addr, uint8_t mux_ch );
+ErrorStatus PDM_init_INA231( _PDM *pdm_ptr, I2C_TypeDef *I2Cx, uint8_t INA231_addr, uint16_t max_current, uint16_t Rshunt, uint8_t i2c_mux_addr, uint8_t mux_ch );
